vMPU: Move files to dedicated folder
