+++ core/vmpu/inc/vmpu_freescale_k64_mem.h
