+++ core/vmpu/src/freescale_k64/vmpu_freescale_k64_mem.c
